Add image configuration option
The following patch proposes to host nix configuration
files under nix folder that offers default configuration
for an image, defaulting to a release image, which would
be plain spectrum. A hardened default configuration will
be proposed in the near future. In case of configuration
collision between the default configuration and config.nix,
the latter will be taken into account.

Signed-off-by: José Pekkarinen <jose.pekkarinen@unikie.com>